Extract jail dataset utils
